[Profiler] Memory profiler part 8: Mark parameters.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87568

Following the pattern of earlier PRs, we use two methods to extract parameters. The primary one is the Python tracer; both nn.Module and optim.Optimizer collect parameters and in most cases that is sufficient. As a fallback we can analyze the data flow graph and deduce likely parameters based on gradient computation and updates.

Parameter identification has a circular interaction with input identification. Inputs are defined as "not part of the core forward-backward-update loop", but we need inputs for the parameter identification fallback to give us a proxy for the forward pass. Thus, we mark parameters from the python tracer which limits which Tensors get marked as inputs. While not necessary, it adds a bit of robustness. (As shown by the strengthening of the input unit tests.)
ghstack-source-id: 173450216

Differential Revision: [D40238619](https://our.internmc.facebook.com/intern/diff/D40238619/)