Simplify linspace decomp and increase its tolerance
This is an interesting one

Since this is an operation that's intrinsically defined on the reals,
we should perform the ops on that dtype always, and just cast to
the desired dtype at the end. This simplifies the decomposition.

Now, I started looking at this one when I started seeing failures on a
test that's added in a later PR. What's going on here is that, by doing
an upcast to a higher dtype and then cast down to integers, sometimes
there's an off-by-one error. I think this is fine, as the decomposition
is more accurate than the original function, which goes in line with
the whole PrimTorch effort.

ghstack-source-id: 4af9ad0328569a30bd6d720ded677f6b59e3c19e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87203